feat: id값 int 타입에서 long으로 변경

+++ app/src/main/java/com/codesoom/assignment/repository/CatToyRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface CatToyRepository extends CrudRepository<CatToy, Long> {
     List<CatToy> findAll();
-    CatToy findById(int id);
+    CatToy findById(long id);
     CatToy save(CatToy catToy);
-    void deleteById(int id);
+    void deleteById(long id);
 }
